feat:show only 2 numbers

+++ get-array-values.h
@@ -1,3 +1,3 @@
 #include <stddef.h>
 
-int *get_array_values(size_t size);
+float *get_array_values(size_t size);
